refactor: Teacher package removal -Moved Teacher package classes to parent User package -removed unused Teacher package -Fixed imports -Added @Primary annotation for SecurityServiceImpl
