+++ src/main/java/pl/simpleascoding/tutoringplatform/user/TeacherService.java
@@ -1,9 +1,9 @@
-package pl.simpleascoding.tutoringplatform.user.teacher;
+package pl.simpleascoding.tutoringplatform.user;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import pl.simpleascoding.tutoringplatform.util.rscp.RscpDTO;
-import pl.simpleascoding.tutoringplatform.user.teacher.dto.SignAsTeacherDTO;
+import pl.simpleascoding.tutoringplatform.user.dto.SignAsTeacherDTO;
 import pl.simpleascoding.tutoringplatform.user.dto.UserDTO;
 
 public interface TeacherService {
