+++ src/main/java/pl/simpleascoding/tutoringplatform/user/dto/SignAsTeacherDTO.java
@@ -1,0 +1,4 @@
+package pl.simpleascoding.tutoringplatform.user.dto;
+
+public record SignAsTeacherDTO(String username) {
+}
