--- src/main/java/pl/simpleascoding/tutoringplatform/user/teacher/dto/SignAsTeacherDTO.java
@@ -1,4 +1,0 @@
-package pl.simpleascoding.tutoringplatform.user.teacher.dto;
-
-public record SignAsTeacherDTO(String username) {
-}
